requirements.txt
streamlit>=1.30.0
google-generativeai>=0.5.0
firebase-admin>=6.4.0
Pillow>=10.0.0
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
plotly>=5.15.0
twilio>=8.0.0
pyzbar>=0.1.9
statsmodels>=0.14.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
